requirements.txt
chroma-hnswlib==0.7.3
chromadb==0.5.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.0
numpy==1.26.4
openai==0.28.0
pandas==2.2.2
PyJWT==2.8.0
PyKakao==0.0.7
PyMySQL==1.1.0
pytz==2024.1
regex==2024.5.10
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
xmltodict==0.13.0